requirements.txt
Jupyter
black
db-sqlite3
matplotlib
pandas
plotly_express
requests
seaborn
SQLAlchemy
nbformat
pytest
apache-airflow